# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "矿山自动化调度系统 API",
    "description": "面向矿山运输调度场景的数字化管理 API，覆盖运单、排队、装车、称重、看板、告警、调度算法等完整业务链路。",
    "version": "0.1.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "开发环境" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/api/v1/queue/waybills/{waybill_id}/join": {
      "post": {
        "summary": "加入队列",
        "tags": ["排队管理"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["driver_id", "pit_id", "arrival_method"],
                "properties": {
                  "driver_id": { "type": "string", "format": "uuid" },
                  "pit_id": { "type": "string", "format": "uuid" },
                  "arrival_method": { "type": "string" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/queue/waybills/{waybill_id}/call-next": {
      "post": {
        "summary": "叫号",
        "tags": ["排队管理"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["operator_id"],
                "properties": {
                  "operator_id": { "type": "string", "format": "uuid" },
                  "reason": { "type": "string" }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}